// Attendance/frontend/src/types/parent.types.ts
export interface ParentRecord {
  id: string;
  registerNumber: string;
  fatherName: string;
  motherName: string;
  fatherPhone: string;
  motherPhone: string;
  guardianPhone: string;
  address: string;
}

export interface ParentFormData {
  fatherName: string;
  motherName: string;
  fatherPhone: string;
  motherPhone: string;
  guardianPhone: string;
  address: string;
}